fix(dotenv): address startup review feedback
